requirements.txt
# ExamRAG v5 — requirements
# Tested on Python 3.10+

# Core RAG stack
streamlit>=1.32.0
faiss-cpu>=1.7.4
sentence-transformers>=2.6.0
numpy>=1.24.0
requests>=2.31.0

# PDF extraction (try PyMuPDF first, fallback to pdfplumber)
PyMuPDF>=1.23.0
pdfplumber>=0.10.0

# Optional: faster tokenisation (chunking uses heuristic fallback if absent)
tiktoken>=0.5.0

# Data
pandas>=2.0.0
